Refactor: 파일 분리, 폴더 구조 변경
