+++ src/app/(routes)/solved/question/hook/use-user-challenges-progress.tsx
