+++ src/app/(routes)/solved/question/page.tsx
@@ -8,7 +8,7 @@ const Page = () => {
       <Heading />
       <main className="mb-10 flex h-auto justify-center gap-6">
         <QuestionList className="mt-2 h-auto w-[690px]" />
-        <ResponseCompletionRate className="h-auto w-[306px]" />
+        <ResponseCompletionRate className="h-[787px] w-[306px]" />
       </main>
     </>
   );
